Issue-5: restructuring
