+++ src/main/java/com/honestefforts/fixengine/service/service/FixEngineService.java
@@ -10,8 +10,8 @@ import com.honestefforts.fixengine.model.validation.ValidationError;
 import com.honestefforts.fixengine.service.config.TagTypeMapConfig;
 import com.honestefforts.fixengine.service.converter.FixConverterFactory;
 import com.honestefforts.fixengine.service.converter.messagetypes.BusinessMessageRejectConverter;
-import com.honestefforts.fixengine.service.validation.BeginStringFixValidator;
-import com.honestefforts.fixengine.service.validation.FixHeaderFixValidator;
+import com.honestefforts.fixengine.service.validation.header.BeginStringFixValidator;
+import com.honestefforts.fixengine.service.validation.header.FixHeaderFixValidator;
 import com.honestefforts.fixengine.service.validation.FixValidatorFactory;
 import java.util.List;
 import java.util.Map;
