+++ src/main/java/com/honestefforts/fixengine/service/validation/header/BeginStringFixValidator.java
@@ -1,4 +1,4 @@
-package com.honestefforts.fixengine.service.validation;
+package com.honestefforts.fixengine.service.validation.header;
 
 import ch.qos.logback.core.util.StringUtil;
 import com.honestefforts.fixengine.model.message.tags.RawTag;
