+++ src/main/java/com/honestefforts/fixengine/service/validation/header/MessageTypeFixValidator.java
@@ -1,4 +1,4 @@
-package com.honestefforts.fixengine.service.validation;
+package com.honestefforts.fixengine.service.validation.header;
 
 import com.honestefforts.fixengine.model.message.tags.RawTag;
 import com.honestefforts.fixengine.model.validation.FixValidator;
